requirements.txt
torch
torchvision
numpy
Pillow
tqdm
